+++ R/Globals-class.R
@@ -50,19 +50,27 @@ Globals <- function(object = list(), ...) {
 as.Globals <- function(x, ...) UseMethod("as.Globals")
 
 #' @export
+as.Globals.default <- function(x, ...) {
+  stopf("Don't know how to coerce a %s to Globals", class(x)[1])
+}
+
+#' @export
 as.Globals.Globals <- function(x, ...) x
 
 #' @export
 as.Globals.list <- function(x, ...) {
-  ## Use the globals environments as the locals?
-  ## (with emptyenv() as the fallback)
-  where <- attr(x, "where", exact = TRUE)
-  if (is.null(where)) {
-    where <- lapply(x, FUN = environment_of)
-    names(where) <- names(x)
-    attr(x, "where") <- where
+  if (length(x) > 0L) {
+    stop_if_not(!is.null(names(x)))
+  
+    ## Use the globals environments as the locals?
+    ## (with emptyenv() as the fallback)
+    where <- attr(x, "where", exact = TRUE)
+    if (is.null(where)) {
+      where <- lapply(x, FUN = environment_of)
+      names(where) <- names(x)
+      attr(x, "where") <- where
+    }
   }
-
   Globals(x, ...)
 }
 
@@ -94,6 +102,7 @@ as.Globals.list <- function(x, ...) {
 
 
 assign_Globals <- function(x, name, value) {
+  stop_if_not(is.character(name), !is.na(name), nchar(name) > 0L)
   where <- attr(x, "where", exact = TRUE)
   stop_if_not(!is.null(where))
   
@@ -114,13 +123,51 @@ assign_Globals <- function(x, name, value) {
     }
   }
 
+  attr(x, "where") <- where
+
   ## Avoid call this function recursively
   class <- class(x)
   class(x) <- NULL
   x[[name]] <- value
   class(x) <- class
 
-  attr(x, "where") <- where
+  invisible(x)
+}
+
+
+#' @export
+`[<-.Globals` <- function(x, names, value) {
+  stop_if_not(
+    length(names) == length(value),
+    is.character(names), !anyNA(names), all(nchar(names) > 0)
+  )
+
+  if (inherits(value, "Globals")) {
+    where <- attr(value, "where")
+  } else if (is.list(value)) {
+    where <- lapply(value, FUN = environment_of)
+  } else {
+    stopf("Unsupported class of 'value': %s", class(value)[1])
+  }
+  stop_if_not(length(where) == length(value))
+
+  x_where <- attr(x, "where", exact = TRUE)
+  stop_if_not(!is.null(x_where))
+
+  class <- class(x)
+  class(x) <- NULL
+  
+  for (kk in seq_along(value)) {
+    name <- names[kk]
+    x[[name]] <- value[[kk]]
+    x_where[[name]] <- where[[kk]]
+  }
+
+  stop_if_not(length(x_where) == length(x))
+
+  attr(x, "where") <- x_where
+  class(x) <- class
+
   invisible(x)
 }
 
@@ -172,8 +219,8 @@ c.Globals <- function(x, ...) {
     x <- c(x, g)
   }
 
-  class(x) <- clazz
   attr(x, "where") <- where
+  class(x) <- clazz
 
   stop_if_not(
     length(where) == length(x),